working example of using `populate()` (using `$in` under the hood) in sample apps
